Remove bindings
