+++ crates/ruff/src/rules/ssort/mod.rs
@@ -1,4 +1,3 @@
-mod bindings;
 mod builtins;
 mod requirements;
 pub(crate) mod rules;
