+++ crates/ruff/src/rules/ssort/requirements.rs
@@ -365,87 +365,47 @@ fn add_arguments_to_bindings<'a>(bindings: &mut IndexSet<&'a str>, arguments: &'
 #[cfg(test)]
 mod tests {
     use super::*;
-    use pretty_assertions::assert_eq;
     use rustpython_parser::lexer::lex;
     use rustpython_parser::{parse_tokens, Mode};
     use unindent::unindent;
 
-    #[test]
-    fn function_def() {
-        let stmt = parse(
-            r#"
+    fn parse(source: &str) -> Stmt {
+        let source = unindent(source);
+        let tokens = lex(&source, Mode::Module);
+        let parsed = parse_tokens(tokens, Mode::Module, "test.py").unwrap();
+        match parsed {
+            Mod::Module(ModModule { body, .. }) => body.into_iter().next().unwrap(),
+            _ => panic!("Unsupported module type"),
+        }
+    }
+
+    pub(super) fn stmt_bindings(stmt: &Stmt) -> Vec<&str> {
+        let mut requirements = AstVisitor::default();
+        requirements.visit_stmt(stmt);
+        Vec::from_iter(requirements.bindings)
+    }
+
+    mod bindings {
+        use super::*;
+        use pretty_assertions::assert_eq;
+
+        #[test]
+        fn function_def() {
+            let stmt = parse(
+                r#"
                 @a
                 def b(c: d = e, /, f: g = h, *i: j, k: l = m, **n: o) -> p:
                     q = r
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "l",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "o",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "p",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "r",
-                    is_deferred: true,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b"]);
+        }
 
-    #[test]
-    fn function_def_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn function_def_with_walrus() {
+            let stmt = parse(
+                r#"
                 @(a := b)
                 def c(
                     d: (e := f) = (g := h),
@@ -457,220 +417,31 @@ mod tests {
                 ) -> (y := z):
                     aa = ab
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "u",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "p",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "s",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "x",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "z",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "ab",
-                    is_deferred: true,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(
+                bindings,
+                ["a", "g", "l", "t", "e", "j", "o", "r", "w", "y", "c"]
+            );
+        }
 
-    #[test]
-    fn function_def_with_bindings() {
-        let stmt = parse(
-            r#"
-                @(a := b)
-                def c(
-                    d: (e := f) = (g := h),
-                    /,
-                    i: (j := k) = (l := m),
-                    *n: (o := p),
-                    q: (r := s) = (t := u),
-                    **v: (w := x)
-                ) -> (y := z):
-                    _ = a, c, d, e, g, i, j, l, n, o, q, r, t, v, w, y
-            "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "u",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "p",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "s",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "x",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "z",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn async_function_def() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn async_function_def() {
+            let stmt = parse(
+                r#"
                 @a
                 async def b(c: d = e, /, f: g = h, *i: j, k: l = m, **n: o) -> p:
                     q = r
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "l",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "o",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "p",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "r",
-                    is_deferred: true,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b"]);
+        }
 
-    #[test]
-    fn async_function_def_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn async_function_def_with_walrus() {
+            let stmt = parse(
+                r#"
                 @(a := b)
                 async def c(
                     d: (e := f) = (g := h),
@@ -682,688 +453,198 @@ mod tests {
                 ) -> (y := z):
                     aa = ab
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "u",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "p",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "s",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "x",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "z",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "ab",
-                    is_deferred: true,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(
+                bindings,
+                ["a", "g", "l", "t", "e", "j", "o", "r", "w", "y", "c"]
+            );
+        }
 
-    #[test]
-    fn async_function_def_with_bindings() {
-        let stmt = parse(
-            r#"
-                @(a := b)
-                async def c(
-                    d: (e := f) = (g := h),
-                    /,
-                    i: (j := k) = (l := m),
-                    *n: (o := p),
-                    q: (r := s) = (t := u),
-                    **v: (w := x)
-                ) -> (y := z):
-                    _ = a, c, d, e, g, i, j, l, n, o, q, r, t, v, w, y
-            "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "u",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "p",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "s",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "x",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "z",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn class_def() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn class_def() {
+            let stmt = parse(
+                r#"
                 @a
                 class b(c, d=f):
                     g = h
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b"]);
+        }
 
-    #[test]
-    fn class_def_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn class_def_with_walrus() {
+            let stmt = parse(
+                r#"
                 @(a := b)
                 class c((d := e), f=(g := h)):
                     i = j
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "d", "g", "c"]);
+        }
 
-    #[test]
-    fn class_def_with_bindings() {
-        let stmt = parse(
-            r#"
-                @(a := b)
-                class c((d := e), f=(g := h)):
-                    _ = a, c, d, g
-            "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn return_() {
+            let stmt = parse(r#"return a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn return_() {
-        let stmt = parse(r#"return a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            }]
-        );
-    }
+        #[test]
+        fn return_with_walrus() {
+            let stmt = parse(r#"return (a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn return_with_walrus() {
-        let stmt = parse(r#"return (a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            }]
-        );
-    }
+        #[test]
+        fn delete() {
+            let stmt = parse(r#"del a, b.c, d[e:f:g]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn delete() {
-        let stmt = parse(r#"del a, b.c, d[e:f:g]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+        #[test]
+        fn delete_with_walrus() {
+            let stmt = parse(r#"del a, (b := c).d, (e := f)[(g := h) : (i := j) : (k := l)]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "e", "g", "i", "k"]);
+        }
 
-    #[test]
-    fn delete_with_walrus() {
-        let stmt = parse(r#"del a, (b := c).d, (e := f)[(g := h) : (i := j) : (k := l)]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "l",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+        #[test]
+        fn assign() {
+            let stmt = parse(r#"a = b, c.d, [e, *f], *g = h"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "b", "e", "f", "g"]);
+        }
 
-    #[test]
-    fn assign() {
-        let stmt = parse(r#"a = b, c.d, [e, *f], *g = h"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+        #[test]
+        fn assign_with_walrus() {
+            let stmt = parse(r#"a = b, (c := d).e, [f, *g], *h = (i := j)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["i", "a", "b", "c", "f", "g", "h"]);
+        }
 
-    #[test]
-    fn assign_with_walrus() {
-        let stmt = parse(r#"a = b, (c := d).e, [f, *g], *h = (i := j)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+        #[test]
+        fn aug_assign() {
+            let stmt = parse(r#"a += b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn aug_assign() {
-        let stmt = parse(r#"a += b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+        #[test]
+        fn aug_assign_with_walrus() {
+            let stmt = parse(r#"a += (b := c)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "a"]);
+        }
 
-    #[test]
-    fn aug_assign_with_walrus() {
-        let stmt = parse(r#"a += (b := c)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                }
-            ]
-        );
-    }
+        #[test]
+        fn ann_assign() {
+            let stmt = parse(r#"a: b = c"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn ann_assign() {
-        let stmt = parse(r#"a: b = c"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn ann_assign_with_walrus() {
+            let stmt = parse(r#"a: (b := c) = (d := e)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["d", "b", "a"]);
+        }
 
-    #[test]
-    fn ann_assign_with_walrus() {
-        let stmt = parse(r#"a: (b := c) = (d := e)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn for_() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn for_() {
+            let stmt = parse(
+                r#"
                 for a in b:
                     c = d
                 else:
                     e = f
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn for_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn for_with_walrus() {
+            let stmt = parse(
+                r#"
                 for a in (b := c):
                     d = e
                 else:
                     f = g
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "a", "d", "f"]);
+        }
 
-    #[test]
-    fn async_for_() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn async_for() {
+            let stmt = parse(
+                r#"
                 async for a in b:
                     c = d
                 else:
                     e = f
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn async_for_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn async_for_with_walrus() {
+            let stmt = parse(
+                r#"
                 async for a in (b := c):
                     d = e
                 else:
                     f = g
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "a", "d", "f"]);
+        }
 
-    #[test]
-    fn while_() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn while_() {
+            let stmt = parse(
+                r#"
                 while a:
                     b = c
                 else:
                     d = e
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "d"]);
+        }
 
-    #[test]
-    fn while_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn while_with_walrus() {
+            let stmt = parse(
+                r#"
                 while (a := b):
                     c = d
                 else:
                     e = f
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn if_() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn if_() {
+            let stmt = parse(
+                r#"
                 if a:
                     b = c
                 elif d:
@@ -1371,44 +652,15 @@ mod tests {
                 else:
                     g = h
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "e", "g"]);
+        }
 
-    #[test]
-    fn if_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn if_with_walrus() {
+            let stmt = parse(
+                r#"
                 if (a := b):
                     c = d
                 elif (e := f):
@@ -1416,210 +668,77 @@ mod tests {
                 else:
                     i = j
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e", "g", "i"]);
+        }
 
-    #[test]
-    fn with() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn with() {
+            let stmt = parse(
+                r#"
                 with a as b, c as (d, e):
                     f = g
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "d", "e", "f"]);
+        }
 
-    #[test]
-    fn with_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn with_with_walrus() {
+            let stmt = parse(
+                r#"
                 with (a := b) as c, (d := e) as (f, g):
                     h = i
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "d", "f", "g", "h"]);
+        }
 
-    #[test]
-    fn async_with() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn async_with() {
+            let stmt = parse(
+                r#"
                 async with a as b, c as (d, e):
                     f = g
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "d", "e", "f"]);
+        }
 
-    #[test]
-    fn async_with_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn async_with_with_walrus() {
+            let stmt = parse(
+                r#"
                 async with (a := b) as c, (d := e) as (f, g):
                     h = i
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "d", "f", "g", "h"]);
+        }
 
-    #[test]
-    fn raise() {
-        let stmt = parse(r#"raise a from b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn raise() {
+            let stmt = parse(r#"raise a from b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn raise_with_walrus() {
-        let stmt = parse(r#"raise (a := b) from (c := d)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn raise_with_walrus() {
+            let stmt = parse(r#"raise (a := b) from (c := d)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c"]);
+        }
 
-    #[test]
-    fn try_() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn try_() {
+            let stmt = parse(
+                r#"
                 try:
                     a = b
                 except c as d:
@@ -1629,44 +748,15 @@ mod tests {
                 finally:
                     i = j
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "d", "e", "g", "i"]);
+        }
 
-    #[test]
-    fn try_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn try_with_walrus() {
+            let stmt = parse(
+                r#"
                 try:
                     a = b
                 except (c := d) as e:
@@ -1676,44 +766,15 @@ mod tests {
                 finally:
                     j = k
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e", "f", "h", "j"]);
+        }
 
-    #[test]
-    fn try_star() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn try_star() {
+            let stmt = parse(
+                r#"
                 try:
                     a = b
                 except* c as d:
@@ -1723,44 +784,15 @@ mod tests {
                 finally:
                     i = j
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "d", "e", "g", "i"]);
+        }
 
-    #[test]
-    fn try_star_with_walrus() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn try_star_with_walrus() {
+            let stmt = parse(
+                r#"
                 try:
                     a = b
                 except* (c := d) as e:
@@ -1770,1524 +802,3444 @@ mod tests {
                 finally:
                     j = k
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e", "f", "h", "j"]);
+        }
 
-    #[test]
-    fn assert() {
-        let stmt = parse(r#"assert a, b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn assert() {
+            let stmt = parse(r#"assert a, b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn assert_with_walrus() {
-        let stmt = parse(r#"assert (a := b), (c := d)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn assert_with_walrus() {
+            let stmt = parse(r#"assert (a := b), (c := d)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c"]);
+        }
 
-    #[test]
-    fn import() {
-        let stmt = parse(r#"import a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn import() {
+            let stmt = parse(r#"import a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn import_with_submodule() {
-        let stmt = parse(r#"import a.b.c"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn import_with_submodule() {
+            let stmt = parse(r#"import a.b.c"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn import_with_alias() {
-        let stmt = parse(r#"import a as b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn import_with_alias() {
+            let stmt = parse(r#"import a as b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b"]);
+        }
 
-    #[test]
-    fn import_from() {
-        let stmt = parse(r#"from a import b, c"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn import_from() {
+            let stmt = parse(r#"from a import b, c"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "c"]);
+        }
 
-    #[test]
-    fn import_from_with_alias() {
-        let stmt = parse(r#"from a import b as c, d as e"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn import_from_with_alias() {
+            let stmt = parse(r#"from a import b as c, d as e"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["c", "e"]);
+        }
 
-    #[test]
-    fn global() {
-        let stmt = parse(r#"global a, b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Global
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Global
-                },
-            ]
-        );
-    }
+        #[test]
+        fn global() {
+            let stmt = parse(r#"global a, b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "b"]);
+        }
 
-    #[test]
-    fn nonlocal() {
-        let stmt = parse(r#"nonlocal a, b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::NonLocal
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::NonLocal
-                },
-            ]
-        );
-    }
+        #[test]
+        fn nonlocal() {
+            let stmt = parse(r#"nonlocal a, b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "b"]);
+        }
 
-    #[test]
-    fn pass() {
-        let stmt = parse(r#"pass"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn pass() {
+            let stmt = parse(r#"pass"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn break_() {
-        let stmt = parse(r#"break"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn break_() {
+            let stmt = parse(r#"break"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn continue_() {
-        let stmt = parse(r#"continue"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn continue_() {
+            let stmt = parse(r#"continue"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn bool_op() {
-        let stmt = parse(r#"a and b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn bool_op() {
+            let stmt = parse(r#"a and b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn bool_op_with_walrus() {
-        let stmt = parse(r#"(a := b) and (c := d)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn bool_op_with_walrus() {
+            let stmt = parse(r#"(a := b) and (c := d)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c"]);
+        }
 
-    #[test]
-    fn named_expr() {
-        let stmt = parse(r#"(a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn named_expr() {
+            let stmt = parse(r#"(a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn bin_op() {
-        let stmt = parse(r#"a + b"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn bin_op() {
+            let stmt = parse(r#"a + b"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn bin_op_with_walrus() {
-        let stmt = parse(r#"(a := b) + (c := d)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn bin_op_with_walrus() {
+            let stmt = parse(r#"(a := b) + (c := d)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c"]);
+        }
 
-    #[test]
-    fn unary_op() {
-        let stmt = parse(r#"-a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn unary_op() {
+            let stmt = parse(r#"-a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn unary_op_with_walrus() {
-        let stmt = parse(r#"-(a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn unary_op_with_walrus() {
+            let stmt = parse(r#"-(a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn lambda() {
-        let stmt = parse(r#"lambda a = b, /, c = d, *e, f = g, **h: i"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: true,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn lambda() {
+            let stmt = parse(r#"lambda a = b, /, c = d, *e, f = g, **h: i"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn lambda_with_walrus() {
-        let stmt =
-            parse(r#"lambda a = (b := c), /, d = (e := f), *g, h = (i := j), **k: (l := m)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "m",
-                    is_deferred: true,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn lambda_with_walrus() {
+            let stmt =
+                parse(r#"lambda a = (b := c), /, d = (e := f), *g, h = (i := j), **k: (l := m)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "e", "i"]);
+        }
 
-    #[test]
-    fn if_exp() {
-        let stmt = parse(r#"a if b else c"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn if_exp() {
+            let stmt = parse(r#"a if b else c"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn if_exp_with_walrus() {
-        let stmt = parse(r#"(a := b) if (c := d) else (e := f)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn if_exp_with_walrus() {
+            let stmt = parse(r#"(a := b) if (c := d) else (e := f)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["c", "a", "e"]);
+        }
 
-    #[test]
-    fn dict() {
-        let stmt = parse(r#"{a: b, **c}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn dict() {
+            let stmt = parse(r#"{a: b, **c}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn dict_with_walrus() {
-        let stmt = parse(r#"{(a := b): (c := d), **(e := f)}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn dict_with_walrus() {
+            let stmt = parse(r#"{(a := b): (c := d), **(e := f)}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn set() {
-        let stmt = parse(r#"{a, *b}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn set() {
+            let stmt = parse(r#"{a, *b}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn set_with_walrus() {
-        let stmt = parse(r#"{(a := b), *(c := d)}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn set_with_walrus() {
+            let stmt = parse(r#"{(a := b), *(c := d)}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c"]);
+        }
 
-    #[test]
-    fn list_comp() {
-        let stmt = parse(r#"[a for b in c if d]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn list_comp() {
+            let stmt = parse(r#"[a for b in c if d]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn list_comp_with_walrus() {
-        let stmt = parse(r#"[(a := b) for c in (d := f) if (g := h)]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn list_comp_with_walrus() {
+            let stmt = parse(r#"[(a := b) for c in (d := f) if (g := h)]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["d", "g", "a"]);
+        }
 
-    #[test]
-    fn set_comp() {
-        let stmt = parse(r#"{a for b in c if d}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn set_comp() {
+            let stmt = parse(r#"{a for b in c if d}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn set_comp_with_walrus() {
-        let stmt = parse(r#"{(a := b) for c in (d := f) if (g := h)}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn set_comp_with_walrus() {
+            let stmt = parse(r#"{(a := b) for c in (d := f) if (g := h)}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["d", "g", "a"]);
+        }
 
-    #[test]
-    fn dict_comp() {
-        let stmt = parse(r#"{a: b for c in d if e}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn dict_comp() {
+            let stmt = parse(r#"{a: b for c in d if e}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn dict_comp_with_walrus() {
-        let stmt = parse(r#"{(a := b): (c := d) for e in (f := g) if (h := i)}"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn dict_comp_with_walrus() {
+            let stmt = parse(r#"{(a := b): (c := d) for e in (f := g) if (h := i)}"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["f", "h", "a", "c"]);
+        }
 
-    #[test]
-    fn generator_exp() {
-        let stmt = parse(r#"(a for b in c if d)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn generator_exp() {
+            let stmt = parse(r#"(a for b in c if d)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn generator_exp_with_walrus() {
-        let stmt = parse(r#"((a := b) for c in (d := f) if (g := h))"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn generator_exp_with_walrus() {
+            let stmt = parse(r#"((a := b) for c in (d := f) if (g := h))"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["d", "g", "a"]);
+        }
 
-    #[test]
-    fn await_() {
-        let stmt = parse(r#"await a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn await_() {
+            let stmt = parse(r#"await a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn await_with_walrus() {
-        let stmt = parse(r#"await (a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn await_with_walrus() {
+            let stmt = parse(r#"await (a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn yield_() {
-        let stmt = parse(r#"yield a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn yield_() {
+            let stmt = parse(r#"yield a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn yield_with_walrus() {
-        let stmt = parse(r#"yield (a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn yield_with_walrus() {
+            let stmt = parse(r#"yield (a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn yield_from() {
-        let stmt = parse(r#"yield from a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn yield_from() {
+            let stmt = parse(r#"yield from a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn yield_from_with_walrus() {
-        let stmt = parse(r#"yield from (a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn yield_from_with_walrus() {
+            let stmt = parse(r#"yield from (a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn compare() {
-        let stmt = parse(r#"a < b < c"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn compare() {
+            let stmt = parse(r#"a < b < c"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn compare_with_walrus() {
-        let stmt = parse(r#"(a := b) < (c := d) < (e := f)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn compare_with_walrus() {
+            let stmt = parse(r#"(a := b) < (c := d) < (e := f)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn call() {
-        let stmt = parse(r#"a(b, *c, d=e, **f)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn call() {
+            let stmt = parse(r#"a(b, *c, d=e, **f)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn call_with_walrus() {
-        let stmt = parse(r#"(a := b)((c := d), *(e := f), g=(h := i), **(j := k))"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "k",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn call_with_walrus() {
+            let stmt = parse(r#"(a := b)((c := d), *(e := f), g=(h := i), **(j := k))"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e", "h", "j"]);
+        }
 
-    #[test]
-    fn formatted_value() {
-        let stmt = parse(r#"f"{a:{b}}""#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn formatted_value() {
+            let stmt = parse(r#"f"{a:{b}}""#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn formatted_value_with_walrus() {
-        let stmt = parse(r#"f"{(a := b):{(c := d)}}""#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn formatted_value_with_walrus() {
+            let stmt = parse(r#"f"{(a := b):{(c := d)}}""#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c"]);
+        }
 
-    #[test]
-    fn joined_str() {
-        let stmt = parse(r#"f"{a:{b}} {c:{d}}""#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn joined_str() {
+            let stmt = parse(r#"f"{a:{b}} {c:{d}}""#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn joined_str_with_walrus() {
-        let stmt = parse(r#"f"{(a := b):{(c := d)}} {(e := f):{(g := h)}}""#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn joined_str_with_walrus() {
+            let stmt = parse(r#"f"{(a := b):{(c := d)}} {(e := f):{(g := h)}}""#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e", "g"]);
+        }
 
-    #[test]
-    fn constant() {
-        let stmt = parse(r#"1"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(requirements, [] as [Requirement; 0]);
-    }
+        #[test]
+        fn constant() {
+            let stmt = parse(r#"1"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn attribute() {
-        let stmt = parse(r#"a.b.c"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn attribute() {
+            let stmt = parse(r#"a.b.c"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn attribute_with_walrus() {
-        let stmt = parse(r#"(a := b).c.d"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn attribute_with_walrus() {
+            let stmt = parse(r#"(a := b).c.d"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn subscript() {
-        let stmt = parse(r#"a[b:c:d]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn subscript() {
+            let stmt = parse(r#"a[b:c:d]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn subscript_with_walrus() {
-        let stmt = parse(r#"(a := b)[(c := d):(e := f):(g := h)]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn subscript_with_walrus() {
+            let stmt = parse(r#"(a := b)[(c := d):(e := f):(g := h)]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e", "g"]);
+        }
 
-    #[test]
-    fn starred() {
-        let stmt = parse(r#"*a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn starred() {
+            let stmt = parse(r#"*a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn starred_with_walrus() {
-        let stmt = parse(r#"*(a := b)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "b",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn starred_with_walrus() {
+            let stmt = parse(r#"*(a := b)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a"]);
+        }
 
-    #[test]
-    fn name() {
-        let stmt = parse(r#"a"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [Requirement {
-                name: "a",
-                is_deferred: false,
-                context: RequirementContext::Local
-            },]
-        );
-    }
+        #[test]
+        fn name() {
+            let stmt = parse(r#"a"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn list() {
-        let stmt = parse(r#"[a, b, *c]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn list() {
+            let stmt = parse(r#"[a, b, *c]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn list_with_walrus() {
-        let stmt = parse(r#"[(a := b), (c := d), *(e := f)]"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn list_with_walrus() {
+            let stmt = parse(r#"[(a := b), (c := d), *(e := f)]"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn tuple() {
-        let stmt = parse(r#"(a, b, *c)"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn tuple() {
+            let stmt = parse(r#"(a, b, *c)"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, [] as [&str; 0]);
+        }
 
-    #[test]
-    fn tuple_with_walrus() {
-        let stmt = parse(r#"((a := b), (c := d), *(e := f))"#);
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "d",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+        #[test]
+        fn tuple_with_walrus() {
+            let stmt = parse(r#"((a := b), (c := d), *(e := f))"#);
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["a", "c", "e"]);
+        }
 
-    #[test]
-    fn match_value() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_value() {
+            let stmt = parse(
+                r#"
                 match a:
                     case "" as b if c:
                         d = e
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "d"]);
+        }
 
-    #[test]
-    fn match_singleton() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_singleton() {
+            let stmt = parse(
+                r#"
                 match a:
                     case None as b if c:
                         d = e
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "d"]);
+        }
 
-    #[test]
-    fn match_sequence() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_sequence() {
+            let stmt = parse(
+                r#"
                 match a:
                     case [b, *c, _] as e if f:
                         g = h
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "f",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "c", "e", "g"]);
+        }
 
-    #[test]
-    fn match_mapping() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_mapping() {
+            let stmt = parse(
+                r#"
                 match a:
                     case {"b": c, "d": _, **e} as f if g:
                         h = i
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "i",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["c", "e", "f", "h"]);
+        }
 
-    #[test]
-    fn match_class() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_class() {
+            let stmt = parse(
+                r#"
                 match a:
                     case b(c, d=e, f=_) as g if h:
                         i=j
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "b",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "h",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "j",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["c", "e", "g", "i"]);
+        }
 
-    #[test]
-    fn match_star() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_star() {
+            let stmt = parse(
+                r#"
                 match a:
                     case [*_] as b if c:
                         d = e
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
-                    name: "a",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "c",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "d"]);
+        }
 
-    #[test]
-    fn match_or() {
-        let stmt = parse(
-            r#"
+        #[test]
+        fn match_or() {
+            let stmt = parse(
+                r#"
                 match a:
                     case [b] | (c) as d if e:
                         f = g
             "#,
-        );
-        let requirements = stmt_requirements(&stmt);
-        assert_eq!(
-            requirements,
-            [
-                Requirement {
+            );
+            let bindings = stmt_bindings(&stmt);
+            assert_eq!(bindings, ["b", "c", "d", "f"]);
+        }
+    }
+
+    mod requirements {
+        use super::*;
+        use pretty_assertions::assert_eq;
+
+        #[test]
+        fn function_def() {
+            let stmt = parse(
+                r#"
+                @a
+                def b(c: d = e, /, f: g = h, *i: j, k: l = m, **n: o) -> p:
+                    q = r
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "l",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "o",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "p",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "r",
+                        is_deferred: true,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn function_def_with_walrus() {
+            let stmt = parse(
+                r#"
+                @(a := b)
+                def c(
+                    d: (e := f) = (g := h),
+                    /,
+                    i: (j := k) = (l := m),
+                    *n: (o := p),
+                    q: (r := s) = (t := u),
+                    **v: (w := x)
+                ) -> (y := z):
+                    aa = ab
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "u",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "p",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "s",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "x",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "z",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "ab",
+                        is_deferred: true,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn function_def_with_bindings() {
+            let stmt = parse(
+                r#"
+                @(a := b)
+                def c(
+                    d: (e := f) = (g := h),
+                    /,
+                    i: (j := k) = (l := m),
+                    *n: (o := p),
+                    q: (r := s) = (t := u),
+                    **v: (w := x)
+                ) -> (y := z):
+                    _ = a, c, d, e, g, i, j, l, n, o, q, r, t, v, w, y
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "u",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "p",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "s",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "x",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "z",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_function_def() {
+            let stmt = parse(
+                r#"
+                @a
+                async def b(c: d = e, /, f: g = h, *i: j, k: l = m, **n: o) -> p:
+                    q = r
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "l",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "o",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "p",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "r",
+                        is_deferred: true,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_function_def_with_walrus() {
+            let stmt = parse(
+                r#"
+                @(a := b)
+                async def c(
+                    d: (e := f) = (g := h),
+                    /,
+                    i: (j := k) = (l := m),
+                    *n: (o := p),
+                    q: (r := s) = (t := u),
+                    **v: (w := x)
+                ) -> (y := z):
+                    aa = ab
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "u",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "p",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "s",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "x",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "z",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "ab",
+                        is_deferred: true,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_function_def_with_bindings() {
+            let stmt = parse(
+                r#"
+                @(a := b)
+                async def c(
+                    d: (e := f) = (g := h),
+                    /,
+                    i: (j := k) = (l := m),
+                    *n: (o := p),
+                    q: (r := s) = (t := u),
+                    **v: (w := x)
+                ) -> (y := z):
+                    _ = a, c, d, e, g, i, j, l, n, o, q, r, t, v, w, y
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "u",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "p",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "s",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "x",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "z",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn class_def() {
+            let stmt = parse(
+                r#"
+                @a
+                class b(c, d=f):
+                    g = h
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn class_def_with_walrus() {
+            let stmt = parse(
+                r#"
+                @(a := b)
+                class c((d := e), f=(g := h)):
+                    i = j
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn class_def_with_bindings() {
+            let stmt = parse(
+                r#"
+                @(a := b)
+                class c((d := e), f=(g := h)):
+                    _ = a, c, d, g
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn return_() {
+            let stmt = parse(r#"return a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
                     name: "a",
                     is_deferred: false,
                     context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "e",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-                Requirement {
-                    name: "g",
-                    is_deferred: false,
-                    context: RequirementContext::Local
-                },
-            ]
-        );
-    }
+                }]
+            );
+        }
 
-    fn parse(source: &str) -> Stmt {
-        let source = unindent(source);
-        let tokens = lex(&source, Mode::Module);
-        let parsed = parse_tokens(tokens, Mode::Module, "test.py").unwrap();
-        match parsed {
-            Mod::Module(ModModule { body, .. }) => body.into_iter().next().unwrap(),
-            _ => panic!("Unsupported module type"),
+        #[test]
+        fn return_with_walrus() {
+            let stmt = parse(r#"return (a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                }]
+            );
+        }
+
+        #[test]
+        fn delete() {
+            let stmt = parse(r#"del a, b.c, d[e:f:g]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn delete_with_walrus() {
+            let stmt = parse(r#"del a, (b := c).d, (e := f)[(g := h) : (i := j) : (k := l)]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "l",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn assign() {
+            let stmt = parse(r#"a = b, c.d, [e, *f], *g = h"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn assign_with_walrus() {
+            let stmt = parse(r#"a = b, (c := d).e, [f, *g], *h = (i := j)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn aug_assign() {
+            let stmt = parse(r#"a += b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn aug_assign_with_walrus() {
+            let stmt = parse(r#"a += (b := c)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    }
+                ]
+            );
+        }
+
+        #[test]
+        fn ann_assign() {
+            let stmt = parse(r#"a: b = c"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn ann_assign_with_walrus() {
+            let stmt = parse(r#"a: (b := c) = (d := e)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn for_() {
+            let stmt = parse(
+                r#"
+                for a in b:
+                    c = d
+                else:
+                    e = f
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn for_with_walrus() {
+            let stmt = parse(
+                r#"
+                for a in (b := c):
+                    d = e
+                else:
+                    f = g
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_for_() {
+            let stmt = parse(
+                r#"
+                async for a in b:
+                    c = d
+                else:
+                    e = f
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_for_with_walrus() {
+            let stmt = parse(
+                r#"
+                async for a in (b := c):
+                    d = e
+                else:
+                    f = g
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn while_() {
+            let stmt = parse(
+                r#"
+                while a:
+                    b = c
+                else:
+                    d = e
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn while_with_walrus() {
+            let stmt = parse(
+                r#"
+                while (a := b):
+                    c = d
+                else:
+                    e = f
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn if_() {
+            let stmt = parse(
+                r#"
+                if a:
+                    b = c
+                elif d:
+                    e = f
+                else:
+                    g = h
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn if_with_walrus() {
+            let stmt = parse(
+                r#"
+                if (a := b):
+                    c = d
+                elif (e := f):
+                    g = h
+                else:
+                    i = j
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn with() {
+            let stmt = parse(
+                r#"
+                with a as b, c as (d, e):
+                    f = g
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn with_with_walrus() {
+            let stmt = parse(
+                r#"
+                with (a := b) as c, (d := e) as (f, g):
+                    h = i
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_with() {
+            let stmt = parse(
+                r#"
+                async with a as b, c as (d, e):
+                    f = g
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn async_with_with_walrus() {
+            let stmt = parse(
+                r#"
+                async with (a := b) as c, (d := e) as (f, g):
+                    h = i
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn raise() {
+            let stmt = parse(r#"raise a from b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn raise_with_walrus() {
+            let stmt = parse(r#"raise (a := b) from (c := d)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn try_() {
+            let stmt = parse(
+                r#"
+                try:
+                    a = b
+                except c as d:
+                    e = f
+                else:
+                    g = h
+                finally:
+                    i = j
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn try_with_walrus() {
+            let stmt = parse(
+                r#"
+                try:
+                    a = b
+                except (c := d) as e:
+                    f = g
+                else:
+                    h = i
+                finally:
+                    j = k
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn try_star() {
+            let stmt = parse(
+                r#"
+                try:
+                    a = b
+                except* c as d:
+                    e = f
+                else:
+                    g = h
+                finally:
+                    i = j
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn try_star_with_walrus() {
+            let stmt = parse(
+                r#"
+                try:
+                    a = b
+                except* (c := d) as e:
+                    f = g
+                else:
+                    h = i
+                finally:
+                    j = k
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn assert() {
+            let stmt = parse(r#"assert a, b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn assert_with_walrus() {
+            let stmt = parse(r#"assert (a := b), (c := d)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn import() {
+            let stmt = parse(r#"import a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn import_with_submodule() {
+            let stmt = parse(r#"import a.b.c"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn import_with_alias() {
+            let stmt = parse(r#"import a as b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn import_from() {
+            let stmt = parse(r#"from a import b, c"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn import_from_with_alias() {
+            let stmt = parse(r#"from a import b as c, d as e"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn global() {
+            let stmt = parse(r#"global a, b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Global
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Global
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn nonlocal() {
+            let stmt = parse(r#"nonlocal a, b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::NonLocal
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::NonLocal
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn pass() {
+            let stmt = parse(r#"pass"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn break_() {
+            let stmt = parse(r#"break"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn continue_() {
+            let stmt = parse(r#"continue"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn bool_op() {
+            let stmt = parse(r#"a and b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn bool_op_with_walrus() {
+            let stmt = parse(r#"(a := b) and (c := d)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn named_expr() {
+            let stmt = parse(r#"(a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn bin_op() {
+            let stmt = parse(r#"a + b"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn bin_op_with_walrus() {
+            let stmt = parse(r#"(a := b) + (c := d)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn unary_op() {
+            let stmt = parse(r#"-a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn unary_op_with_walrus() {
+            let stmt = parse(r#"-(a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn lambda() {
+            let stmt = parse(r#"lambda a = b, /, c = d, *e, f = g, **h: i"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: true,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn lambda_with_walrus() {
+            let stmt =
+                parse(r#"lambda a = (b := c), /, d = (e := f), *g, h = (i := j), **k: (l := m)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "m",
+                        is_deferred: true,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn if_exp() {
+            let stmt = parse(r#"a if b else c"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn if_exp_with_walrus() {
+            let stmt = parse(r#"(a := b) if (c := d) else (e := f)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn dict() {
+            let stmt = parse(r#"{a: b, **c}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn dict_with_walrus() {
+            let stmt = parse(r#"{(a := b): (c := d), **(e := f)}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn set() {
+            let stmt = parse(r#"{a, *b}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn set_with_walrus() {
+            let stmt = parse(r#"{(a := b), *(c := d)}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn list_comp() {
+            let stmt = parse(r#"[a for b in c if d]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn list_comp_with_walrus() {
+            let stmt = parse(r#"[(a := b) for c in (d := f) if (g := h)]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn set_comp() {
+            let stmt = parse(r#"{a for b in c if d}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn set_comp_with_walrus() {
+            let stmt = parse(r#"{(a := b) for c in (d := f) if (g := h)}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn dict_comp() {
+            let stmt = parse(r#"{a: b for c in d if e}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn dict_comp_with_walrus() {
+            let stmt = parse(r#"{(a := b): (c := d) for e in (f := g) if (h := i)}"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn generator_exp() {
+            let stmt = parse(r#"(a for b in c if d)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn generator_exp_with_walrus() {
+            let stmt = parse(r#"((a := b) for c in (d := f) if (g := h))"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn await_() {
+            let stmt = parse(r#"await a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn await_with_walrus() {
+            let stmt = parse(r#"await (a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn yield_() {
+            let stmt = parse(r#"yield a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn yield_with_walrus() {
+            let stmt = parse(r#"yield (a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn yield_from() {
+            let stmt = parse(r#"yield from a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn yield_from_with_walrus() {
+            let stmt = parse(r#"yield from (a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn compare() {
+            let stmt = parse(r#"a < b < c"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn compare_with_walrus() {
+            let stmt = parse(r#"(a := b) < (c := d) < (e := f)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn call() {
+            let stmt = parse(r#"a(b, *c, d=e, **f)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn call_with_walrus() {
+            let stmt = parse(r#"(a := b)((c := d), *(e := f), g=(h := i), **(j := k))"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "k",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn formatted_value() {
+            let stmt = parse(r#"f"{a:{b}}""#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn formatted_value_with_walrus() {
+            let stmt = parse(r#"f"{(a := b):{(c := d)}}""#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn joined_str() {
+            let stmt = parse(r#"f"{a:{b}} {c:{d}}""#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn joined_str_with_walrus() {
+            let stmt = parse(r#"f"{(a := b):{(c := d)}} {(e := f):{(g := h)}}""#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn constant() {
+            let stmt = parse(r#"1"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(requirements, [] as [Requirement; 0]);
+        }
+
+        #[test]
+        fn attribute() {
+            let stmt = parse(r#"a.b.c"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn attribute_with_walrus() {
+            let stmt = parse(r#"(a := b).c.d"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn subscript() {
+            let stmt = parse(r#"a[b:c:d]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn subscript_with_walrus() {
+            let stmt = parse(r#"(a := b)[(c := d):(e := f):(g := h)]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn starred() {
+            let stmt = parse(r#"*a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn starred_with_walrus() {
+            let stmt = parse(r#"*(a := b)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "b",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn name() {
+            let stmt = parse(r#"a"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [Requirement {
+                    name: "a",
+                    is_deferred: false,
+                    context: RequirementContext::Local
+                },]
+            );
+        }
+
+        #[test]
+        fn list() {
+            let stmt = parse(r#"[a, b, *c]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn list_with_walrus() {
+            let stmt = parse(r#"[(a := b), (c := d), *(e := f)]"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn tuple() {
+            let stmt = parse(r#"(a, b, *c)"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn tuple_with_walrus() {
+            let stmt = parse(r#"((a := b), (c := d), *(e := f))"#);
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "d",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_value() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case "" as b if c:
+                        d = e
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_singleton() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case None as b if c:
+                        d = e
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_sequence() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case [b, *c, _] as e if f:
+                        g = h
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "f",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_mapping() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case {"b": c, "d": _, **e} as f if g:
+                        h = i
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "i",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_class() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case b(c, d=e, f=_) as g if h:
+                        i=j
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "b",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "h",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "j",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_star() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case [*_] as b if c:
+                        d = e
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "c",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
+        }
+
+        #[test]
+        fn match_or() {
+            let stmt = parse(
+                r#"
+                match a:
+                    case [b] | (c) as d if e:
+                        f = g
+            "#,
+            );
+            let requirements = stmt_requirements(&stmt);
+            assert_eq!(
+                requirements,
+                [
+                    Requirement {
+                        name: "a",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "e",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                    Requirement {
+                        name: "g",
+                        is_deferred: false,
+                        context: RequirementContext::Local
+                    },
+                ]
+            );
         }
     }
 }
